+++ src/main/java/com/dope/breaking/security/config/WebMvcConfig.java
@@ -1,4 +1,4 @@
-package com.dope.breaking.Security.config;
+package com.dope.breaking.security.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
